+++ index.d.ts
@@ -1,5 +1,5 @@
 export { Canvas } from "./lib/canvas.js";
 export { Colors } from "./lib/colors.js";
 export { Path } from "./lib/path.js";
-export { Position, RGBAColor, CanvasOptions, Resolution, Key, PixelFormat } from "./lib/types.js";
+export { Position, RGBAColor, CanvasOptions, Resolution, Key, PixelFormat, PixelFormats } from "./lib/types.js";
 export * from "./lib/sdl2int.js";